move various ToJS functions to to_js.go
